E2E-AD/contrastive-eeg-video-finetuning/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==24.2.0
black==23.7.0
blinker==1.9.0
braceexpand==0.1.7
cachetools==5.5.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.7
cmake==3.31.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
deepspeed==0.16.0
docker-pycreds==0.4.0
einops==0.8.0
et_xmlfile==2.0.0
fairscale==0.4.13
filelock==3.13.1
fire==0.7.0
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
git-filter-repo==2.47.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.69.0
h5py==3.12.1
hjson==3.1.0
huggingface-hub==0.26.3
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
intel-cmplr-lib-ur==2025.1.0
intel-openmp==2025.1.0
invisible-watermark==0.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kornia==0.6.9
lazy_loader==0.4
lightning-utilities==0.11.9
lit==18.1.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
mkl==2025.1.0
mkl-service==2.4.0
mne==1.9.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.15.2
natsort==8.4.0
networkx==3.3
ninja==1.11.1.2
numpy==1.24.4
nvidia-ml-py==12.560.30
omegaconf==2.3.0
open_clip_torch==2.29.0
opencv-python==4.6.0.66
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.0
pudb==2024.1.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.1
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.28.1
rich==13.9.4
rpds-py==0.22.1
safetensors==0.4.5
scipy==1.14.1
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
streamlit==1.40.2
sympy==1.13.1
tbb==2022.1.0
tcmlib==1.3.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6
termcolor==2.5.0
timm==1.0.12
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchdata==0.6.1
torchmetrics==1.6.0
torchvision==0.15.2+cu118
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
umf==0.10.0
urllib3==1.26.20
urwid==2.6.16
urwid_readline==0.15.1
wandb==0.18.7
watchdog==6.0.0
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.1.3
xformers==0.0.22
yarl==1.18.3
